requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
keras
flask
streamlit
pillow
requests